save vfsmount for path extraction to trigger automount
When decoding snapshot file handly it may happen that the snapshot pointed
by the handly currently not mounted, consequently snapshot automount should
be triggered.

In order to trigger automount path_lookup(<full path to snapshot>) need to be
issued. The <full path to snapshot> should be something like:
<zfs mount point>/<ctl dir name>/snapshot/<snapshot name>

Unfortunately, <zfs mount point> part of the path isn't available in the
context of file handly decoding.
A simple way to have a <zfs mount point> in the context of file handly
decoding is to save pointer to struct vfsmount in the context of super block
creation.

Limitations:
	1 Unfortunately it's kinda hack; in general struct vfsmount shouldn't be
		accessed in the context of the filesystem itself.
	2 Once HAVE_MOUNT_NODEV enabled current implementation wont work at all;
		It's looks like possible to port it to kernel version where
		HAVE_MOUNT_NODEV enabled but it will remain rather HACK than
		solution.
	3 Theoretically the proper solution of the problem is to eliminate usage
		of automount and/or mount at all; just generate all snapshot
		related in core structures (inodes/dentries etc.) on access.

Signed-off-by: Andrey Vesnovaty <andreyv@infinidat.com>

Conflicts:
	include/sys/zfs_vfsops.h
	module/zfs/zfs_vfsops.c

Conflicts:

	include/sys/zfs_vfsops.h
	module/zfs/zfs_ctldir.c
	module/zfs/zfs_vfsops.c

Conflicts:
	module/zfs/zfs_vfsops.c